Project 1: Create 5 LED diodes representing 5 bits binary numbers.
This is a very small arduino project that I learnt and applied in my way.

Main Idea: We have 5 LED lamps, each lamp represents a bit in binary numbering system. They light at the sequence from 0 to 31.

Example: At decimal number 25, binary is 11001. Therefore, the lamps number {1, 4, 5} are lit for a second then it repeats for 26...